Move test sentence generation to script
